clean evaluate_r1cs_matrix_mles
