+++ jolt-core/src/r1cs/mod.rs
@@ -5,6 +5,7 @@ pub mod jolt_constraints;
 pub mod key;
 pub mod ops;
 pub mod spartan;
+pub mod special_polys;
 
 #[cfg(test)]
 pub mod test;
